+++ pkg/store/proxy.go
@@ -286,7 +286,7 @@ func (s *ProxyStore) TSDBInfos() []infopb.TSDBInfo {
 
 func (s *ProxyStore) Series(originalRequest *storepb.SeriesRequest, srv storepb.Store_SeriesServer) error {
 	// TODO(bwplotka): This should be part of request logger, otherwise it does not make much sense. Also, could be
-	// tiggered by tracing span to reduce cognitive load.
+	// triggered by tracing span to reduce cognitive load.
 	reqLogger := log.With(s.logger, "component", "proxy")
 	if s.debugLogging {
 		reqLogger = log.With(reqLogger, "request", originalRequest.String())
@@ -327,6 +327,7 @@ func (s *ProxyStore) Series(originalRequest *storepb.SeriesRequest, srv storepb.
 	r := &storepb.SeriesRequest{
 		MinTime:                 originalRequest.MinTime,
 		MaxTime:                 originalRequest.MaxTime,
+		Limit:                   originalRequest.Limit,
 		Matchers:                append(storeMatchers, MatchersForLabelSets(storeLabelSets)...),
 		Aggregates:              originalRequest.Aggregates,
 		MaxResolutionWindow:     originalRequest.MaxResolutionWindow,
@@ -363,7 +364,13 @@ func (s *ProxyStore) Series(originalRequest *storepb.SeriesRequest, srv storepb.
 	level.Debug(reqLogger).Log("msg", "Series: started fanout streams", "status", strings.Join(storeDebugMsgs, ";"))
 
 	respHeap := NewResponseDeduplicator(NewProxyResponseLoserTree(storeResponses...))
+
+	i := 0
 	for respHeap.Next() {
+		i++
+		if r.Limit > 0 && i > int(r.Limit) {
+			break
+		}
 		resp := respHeap.At()
 
 		if resp.GetWarning() != "" && (r.PartialResponseDisabled || r.PartialResponseStrategy == storepb.PartialResponseStrategy_ABORT) {
@@ -381,7 +388,7 @@ func (s *ProxyStore) Series(originalRequest *storepb.SeriesRequest, srv storepb.
 // LabelNames returns all known label names.
 func (s *ProxyStore) LabelNames(ctx context.Context, originalRequest *storepb.LabelNamesRequest) (*storepb.LabelNamesResponse, error) {
 	// TODO(bwplotka): This should be part of request logger, otherwise it does not make much sense. Also, could be
-	// tiggered by tracing span to reduce cognitive load.
+	// triggered by tracing span to reduce cognitive load.
 	reqLogger := log.With(s.logger, "component", "proxy")
 	if s.debugLogging {
 		reqLogger = log.With(reqLogger, "request", originalRequest.String())
@@ -419,6 +426,7 @@ func (s *ProxyStore) LabelNames(ctx context.Context, originalRequest *storepb.La
 		End:                     originalRequest.End,
 		Matchers:                append(storeMatchers, MatchersForLabelSets(storeLabelSets)...),
 		WithoutReplicaLabels:    originalRequest.WithoutReplicaLabels,
+		Hints:                   originalRequest.Hints,
 	}
 
 	var (
@@ -465,8 +473,13 @@ func (s *ProxyStore) LabelNames(ctx context.Context, originalRequest *storepb.La
 		return nil, err
 	}
 
+	result := strutil.MergeUnsortedSlices(names...)
+	if originalRequest.Limit > 0 && len(result) > int(originalRequest.Limit) {
+		result = result[:originalRequest.Limit]
+	}
+
 	return &storepb.LabelNamesResponse{
-		Names:    strutil.MergeUnsortedSlices(names...),
+		Names:    result,
 		Warnings: warnings,
 	}, nil
 }
@@ -476,7 +489,7 @@ func (s *ProxyStore) LabelValues(ctx context.Context, originalRequest *storepb.L
 	*storepb.LabelValuesResponse, error,
 ) {
 	// TODO(bwplotka): This should be part of request logger, otherwise it does not make much sense. Also, could be
-	// tiggered by tracing span to reduce cognitive load.
+	// triggered by tracing span to reduce cognitive load.
 	reqLogger := log.With(s.logger, "component", "proxy")
 	if s.debugLogging {
 		reqLogger = log.With(reqLogger, "request", originalRequest.String())
@@ -520,6 +533,7 @@ func (s *ProxyStore) LabelValues(ctx context.Context, originalRequest *storepb.L
 		End:                     originalRequest.End,
 		Matchers:                append(storeMatchers, MatchersForLabelSets(storeLabelSets)...),
 		WithoutReplicaLabels:    originalRequest.WithoutReplicaLabels,
+		Limit:                   originalRequest.Limit,
 	}
 
 	var (
@@ -567,8 +581,13 @@ func (s *ProxyStore) LabelValues(ctx context.Context, originalRequest *storepb.L
 		return nil, err
 	}
 
+	vals := strutil.MergeUnsortedSlices(all...)
+	if originalRequest.Limit > 0 && len(vals) > int(originalRequest.Limit) {
+		vals = vals[:originalRequest.Limit]
+	}
+
 	return &storepb.LabelValuesResponse{
-		Values:   strutil.MergeUnsortedSlices(all...),
+		Values:   vals,
 		Warnings: warnings,
 	}, nil
 }